// Frontend/client/model/invoiceDto.ts
/**
 * BillingSoftware.Web
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DocumentInformationDto } from './documentInformationDto';
import { Status } from './status';

export interface InvoiceDto { 
    invoiceNumber?: string;
    invoiceDate?: Date;
    paymentTerm?: Date;
    status?: Status;
    subject?: string;
    headerText?: string;
    flowText?: string;
    documentInformation?: DocumentInformationDto;
}